Desktop/Scorpius1.2-main/packages/core/pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = "Core utilities and orchestration for the Scorpius Enterprise Platform"
authors = ["Scorpius Team <team@scorpius.dev>"]
license = "MIT"
readme = "README.md"
packages = [{include = "core"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
structlog = "^23.2.0"
prometheus-client = "^0.19.0"
httpx = "^0.25.2"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["--strict-markers", "--cov=core"]
